Removed executable portion of rust and implemented as a library package to be published to crates.io. Added proper sections to Cargo.toml. Added section in the ReadMe dedicated to rust usage. Modified name to recursive_matching.
