+++ README.md
@@ -66,7 +66,7 @@ For more information, see `/python/demo.ipynb`
 # Changelog
 
 * Feb 06, 2025 [v1.0]: First release - python implementation.
-* Mar 02, 2025 [v1.1]: Rust implementation, doc fixes, python file name changes.
+* Mar 02, 2025 [v1.0.1]: Rust implementation, doc fixes, python file name changes.
 
 # Modules
 The algorithm will be implemented in three languages: Python, Rust, C.
@@ -104,7 +104,39 @@ print(f"{matches=}")
 ```
 
 ## Rust
-TBA.
+This implementation can be found under `\rust`. To use the crate, add the
+following to your `Cargo.toml`.
+
+```toml
+[dependencies]
+recursive_matching = "1.0.1"
+```
+
+Next import the crate.
+
+```rust
+use recursive_matching::recursive_match;
+```
+
+```rust
+use ndarray::{array, Array2};
+
+let mut matrix: Array2<f32> = array![
+    [0.0, 0.0, 0.0, 0.0, 0.0,],
+    [0.20689655, 0.07407407, 0.04761905, 0.0, 0.23076923],
+    [0.0, 0.0, 0.38461538, 0.0, 0.0],
+    [0.0, 0.0, 0.04347826, 0.5, 0.0],
+    [0.5, 0.0, 0.0, 0.0, 1.0]
+];
+
+let matches = recursive_match(&mut matrix, 1 as usize, true, false);
+
+println!("Matches: {:?}", matches);
+```
 
 ## C
 TBA.
+
+# License
+
+This project is licensed under the GPL v3.0 License.
