+++ rust/matching/Cargo.toml
@@ -1,7 +1,15 @@
 [package]
-name = "matching"
-version = "0.1.0"
+name = "recursive_matching"
+version = "1.0.1"
 edition = "2021"
+authors = ["John Santos johnmarisantos@protonmail.com"]
+description = "Formulating unique assignments recursively."
+license = "GPL 3.0"
+repository = "https://github.com/JohnMVSantos/Recursive-Matching/"
+homepage = "https://github.com/JohnMVSantos/Recursive-Matching/"
+readme = "../../README.md"
+keywords = ["matching", "recursive", "assignments", "rust"]
+categories = ["algorithms"]
 
 [dependencies]
 ndarray = "0.16.1"
